requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
pymupdf>=1.24.0
groq>=0.9.0
google-api-python-client>=2.130.0
google-auth>=2.29.0
python-dotenv>=1.0.1
pydantic>=2.7.0
